보고서 과제 완

+++ 보고서_과제/Q5.cpp
@@ -5,7 +5,7 @@ using namespace std;
 
 int main()
 {
-    int a = 0, b = 0, c = 0;
+    float a = 0, b = 0, c = 0;
     float i = 0, j = 0; 
     
     while(true){
